Item page now actually working lol
Removed all the useRef and useEffect stuff to try to show user and category on component mount because it was not working.

Changed the response from the server so it now returns a flat object with just the user name and category rather than a user object and a category object.

Fixed some lack of error handling on the backend.